rag-document-assistant/requirements.txt
streamlit
flask
langchain
langchain-community
langchain-google-genai
google-generativeai
faiss-cpu
pypdf
nest_asyncio
